# Parse Durations much faster