Without using a regexp.

This shows up in profiles of Prometheus service discovery, since `ParseDuration` is called twice per discovered target.

```
name             old time/op    new time/op    delta
ParseDuration-4    2.54µs ±28%    0.04µs ± 3%   -98.30%  (p=0.008 n=5+5)

name             old alloc/op   new alloc/op   delta
ParseDuration-4      480B ± 0%        0B       -100.00%  (p=0.079 n=4+5)

name             old allocs/op  new allocs/op  delta
ParseDuration-4      2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```

Code inspired by Go standard library `time.ParseDuration`, but simplified and supporting different units.