@bboreham 2023-03-07
Now I realise this makes one of the Prometheus tests fail due to changed error message:

```
--- FAIL: TestPopulateLabels (0.00s)
    manager_test.go:436: 
                Error Trace:    /home/vagrant/src/github.com/prometheus/prometheus/scrape/manager_test.go:436
                Error:          Error message not equal:
                                expected: "error parsing scrape interval: not a valid duration string: \"2notseconds\""
                                actual  : "error parsing scrape interval: unknown unit \"notseconds\" in duration \"2notseconds\""
                Test:           TestPopulateLabels
```

Do you think it's important to retain consistency of error messages?  I.e. shall I fix it here or in the test?