@soundvibe 2023-03-27
This also fails to parse `"0"` durations, breaking a lot of code. Would be great if such improvements would be compatible with previous versions.